helpdesk/requirements.txt
asgiref==3.11.0
click==8.3.1
colorama==0.4.6
dj-database-url==3.1.0
Django==6.0.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==23.0.0
h11==0.16.0
packaging==25.0
psycopg2==2.9.11
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
uvicorn==0.40.0
waitress==3.0.2
whitenoise==6.11.0
